pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docx-processor"
version = "2.0.0-alpha"
description = "Python application to convert Word documents into analyzable formats with intelligent chunking for AI processing"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mammoth>=1.5.0",
    "beautifulsoup4>=4.11.0",
    "Pillow>=9.0.0",
    "lxml>=4.9.0",
    "tiktoken>=0.5.0",
]

[project.scripts]
docx-processor = "docx_processor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["docx_processor"]
